Inline functions refactor + documentation update
Having used the FusionVector*() functions in my code, I have experienced the following problem: if multiple inline functions are used nested in each other, the code fails because the temporary variables have the same name (my environment is: MPLAB X IDE, C30, PIC24F).
On the other hand, the comments for FusionAhrsUpdate are not 100% accurate, the magnetometer values should be in uT because this is the unit for the abandonment check.